Handle instruction size < 5 in API code
Fix for the case in which the instruction's size of the original API
is less then the size of the inconditional JMP to the trampoline.
The idea is to overwrite also the following instrucion(s) and instrument
also them in place---if they are MEMWR as well.
New trampline snippets are generated chained one to the other; only the
last one returns the control to the code.

Signed-off-by: Davide Cingolani <cingodvd@gmail.com>